+++ core/designsystem/src/main/java/com/example/designsystem/component/tobbar/TobAppBar.kt
@@ -1,4 +1,4 @@
-package com.example.designsystem.theme.component
+package com.example.designsystem.component.tobbar
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.layout.Column
@@ -15,10 +15,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.dplay.designsystem.R
-import com.example.designsystem.theme.icon.DplayBaseIcon
-import com.example.designsystem.theme.icon.DplayClickableIcon
-import com.example.designsystem.theme.theme.DPlayTheme
-import com.example.designsystem.theme.util.noRippleClickable
+import com.example.designsystem.util.icon.DplayBaseIcon
+import com.example.designsystem.util.icon.DplayClickableIcon
+import com.example.designsystem.theme.DPlayTheme
+import com.example.designsystem.util.noRippleClickable
 
 @Composable
 fun DplayTopAppBar(
