+++ feature/main/src/main/java/com/example/main/MainActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.designsystem.theme.theme.DPlayTheme
+import com.example.designsystem.theme.DPlayTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
